+++ MonsterTCG/IElement.cs
@@ -7,6 +7,6 @@ namespace MonsterTCG
 {
     public interface IElement
     {
-        int GetDamage(ICard enemy, int damage);
+        double GetDamage(ICard enemy, double damage);
     }
 }
